Remove unused code from demo community
